+++ in-memory/store.test.ts
@@ -4,5 +4,6 @@ import * as store from "./mod.ts";
 Deno.test("@storage/in-memory", async (t) => {
   await testStore(t, store, {
     urlIncludes: "in-memory",
+    orderedByKey: true,
   });
 });